ai: missed some renames, here they are
