+++ Backend/api_v2/app/api/routes/auth_routes.py
@@ -5,7 +5,7 @@ from datetime import timedelta
 from fastapi import APIRouter, Depends, HTTPException, status
 from fastapi.security import OAuth2PasswordRequestForm
 
-from ...models.user import UserCreate, UserPublic
+from ...models.user_models import UserCreate, UserPublic
 from ...services.user_service import UserService
 from ...core.security import create_access_token
 from ...core.deps import get_user_service
